# [stdlib] Use memcpy in `List._realloc` for x13.2 speedups in some benchmarks