--- a/stdlib/src/builtin/reversed.mojo
+++ b/stdlib/src/builtin/reversed.mojo
@@ -78,7 +78,9 @@ fn reversed[T: ReversibleRange](value: T) -> _StridedRange:
 
 fn reversed[
     T: CollectionElement
-](ref [_]value: List[T]) -> _ListIter[T, __lifetime_of(value), False]:
+](ref [_]value: List[T, *_]) -> _ListIter[
+    T, __type_of(value).hint_trivial_type, __lifetime_of(value), False
+]:
     """Get a reversed iterator of the input list.
 
     **Note**: iterators are currently non-raising.