--- a/stdlib/src/builtin/string.mojo
+++ b/stdlib/src/builtin/string.mojo
@@ -691,7 +691,7 @@ struct String(
     """Represents a mutable string."""
 
     # Fields
-    alias _buffer_type = List[UInt8]
+    alias _buffer_type = List[UInt8, hint_trivial_type=True]
     var _buffer: Self._buffer_type
     """The underlying storage for the string."""
 
@@ -715,7 +715,7 @@ struct String(
     # ===------------------------------------------------------------------=== #
 
     @always_inline
-    fn __init__(inout self, owned impl: List[UInt8]):
+    fn __init__(inout self, owned impl: List[UInt8, *_]):
         """Construct a string from a buffer of bytes.
 
         The buffer must be terminated with a null byte:
@@ -735,7 +735,13 @@ struct String(
             impl[-1] == 0,
             "expected last element of String buffer to be null terminator",
         )
-        self._buffer = impl^
+        var size = impl.size
+        var capacity = impl.capacity
+        var ptr = impl.data
+        impl.data = UnsafePointer[UInt8]()
+        self._buffer = Self._buffer_type(
+            unsafe_pointer=ptr, size=size, capacity=capacity
+        )
 
     @always_inline
     fn __init__(inout self):
@@ -1336,7 +1342,7 @@ struct String(
         """
         return self.unsafe_ptr().bitcast[C_char]()
 
-    fn as_bytes(self) -> List[UInt8]:
+    fn as_bytes(self) -> Self._buffer_type:
         """Retrieves the underlying byte sequence encoding the characters in
         this string.
 
@@ -1665,7 +1671,7 @@ struct String(
         var old_len = old.byte_length()
         var new_len = new.byte_length()
 
-        var res = List[UInt8]()
+        var res = Self._buffer_type()
         res.reserve(self_len + (old_len - new_len) * occurrences + 1)
 
         for _ in range(occurrences):
@@ -1795,7 +1801,7 @@ struct String(
         return hash(self._strref_dangerous())
 
     fn _interleave(self, val: String) -> String:
-        var res = List[UInt8]()
+        var res = Self._buffer_type()
         var val_ptr = val.unsafe_ptr()
         var self_ptr = self.unsafe_ptr()
         res.reserve(val.byte_length() * self.byte_length() + 1)
@@ -2157,7 +2163,7 @@ struct String(
             len(fillchar) == 1, "fill char needs to be a one byte literal"
         )
         var fillbyte = fillchar.as_bytes_slice()[0]
-        var buffer = List[UInt8](capacity=width + 1)
+        var buffer = Self._buffer_type(capacity=width + 1)
         buffer.resize(width, fillbyte)
         buffer.append(0)
         memcpy(buffer.unsafe_ptr().offset(start), self.unsafe_ptr(), len(self))