--- a/stdlib/src/collections/list.mojo
+++ b/stdlib/src/collections/list.mojo
@@ -22,7 +22,7 @@ from collections import List
 
 from sys.intrinsics import _type_is_eq
 from os import abort
-from memory import Reference, UnsafePointer
+from memory import Reference, UnsafePointer, memcpy
 from utils import Span
 
 from .optional import Optional