@JoeLoser 2024-08-07
FYI this seems to be hitting some type checking errors (our favorite friend like the error message you mentioned you hit) when I synced this internally.

```
error: cannot implicitly convert 'List[SIMD[uint8, 1], 1]' value to 'List[SIMD[uint8, 1], 0]'
```

We'll need to dig into it.  Did you have a workaround when you previously ran into this?