@gabrieldemarmiesse 2024-08-07
When I got this error (and you have a similar one) I noticed the error is valid and not just a compiler quirk. It's just that the error message is not very clear. You can read it like this:

```
error: cannot implicitly convert 'List[SIMD[uint8, 1], hint_trivial_type=True]' value to 'List[SIMD[uint8, 1], hint_trivial_type=False]'
```
and it makes a lot more sense.

Concerning the trait, it's something we can try in the future. Hopefully, we won't hit any compiler bug.